+++ CHANGELOG.md
@@ -6,6 +6,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [17.2.0] - 2021-01-14
 ### Added
  - Vaddio AV Bridge device driver
  - Added a ContentInput property for VideoConferenceRoutControls